+++ source/robot_model/src/Model.cpp
@@ -661,6 +661,32 @@ Eigen::VectorXd Model::clamp_in_range(const Eigen::VectorXd& vector,
   return lower_limits.cwiseMax(upper_limits.cwiseMin(vector));
 }
 
+state_representation::JointState Model::clamp_in_range(
+    const state_representation::JointState& joint_state,
+    const state_representation::JointStateVariable& state_variable_type) const {
+  using namespace state_representation;
+  Eigen::VectorXd clamped_vector;
+  switch (state_variable_type) {
+    case JointStateVariable::POSITIONS:
+      clamped_vector = this->clamp_in_range(
+          joint_state.get_positions(), this->robot_model_.lowerPositionLimit, this->robot_model_.upperPositionLimit);
+      break;
+    case JointStateVariable::VELOCITIES:
+      clamped_vector = this->clamp_in_range(
+          joint_state.get_velocities(), -this->robot_model_.velocityLimit, this->robot_model_.velocityLimit);
+      break;
+    case JointStateVariable::TORQUES:
+      clamped_vector = this->clamp_in_range(
+          joint_state.get_torques(), -this->robot_model_.effortLimit, this->robot_model_.effortLimit);
+      break;
+    default:
+      return joint_state;
+  }
+  state_representation::JointState joint_state_clamped(joint_state);
+  joint_state_clamped.set_state_variable(clamped_vector, state_variable_type);
+  return joint_state_clamped;
+}
+
 state_representation::JointState Model::clamp_in_range(const state_representation::JointState& joint_state) const {
   state_representation::JointState joint_state_clamped(joint_state);
   joint_state_clamped.set_positions(this->clamp_in_range(joint_state.get_positions(),